+++ services/web/server/src/simcore_service_webserver/catalog/plugin.py
@@ -8,7 +8,6 @@ from pint import UnitRegistry
 from servicelib.aiohttp.application_setup import ModuleCategory, app_module_setup
 
 from . import _handlers
-from ._handlers_reverse_proxy import reverse_proxy_handler
 
 _logger = logging.getLogger(__name__)
 
@@ -22,37 +21,12 @@ _logger = logging.getLogger(__name__)
 )
 def setup_catalog(app: web.Application):
     # ensures routes are names that corresponds to function names
-    for route_def in _handlers.routes:
-        route_def.kwargs["name"] = route_def.handler.__name__
+    assert all(  # nosec
+        route_def.kwargs["name"] == route_def.handler.__name__
+        for route_def in _handlers.routes
+    )
 
     app.add_routes(_handlers.routes)
-
-    # reverse proxy to catalog's API
-    # bind the rest routes with the reverse-proxy-handler
-    app.router.add_routes(
-        [
-            web.get(
-                path="/v0/catalog/dags",
-                handler=reverse_proxy_handler,
-                name="list_catalog_dags",
-            ),
-            web.post(
-                path="/v0/catalog/dags",
-                handler=reverse_proxy_handler,
-                name="create_catalog_dag",
-            ),
-            web.put(
-                path="/v0/catalog/dags/{dag_id}",
-                handler=reverse_proxy_handler,
-                name="replace_catalog_dag",
-            ),
-            web.delete(
-                path="/v0/catalog/dags/{dag_id}",
-                handler=reverse_proxy_handler,
-                name="delete_catalog_dag",
-            ),
-        ]
-    )
 
     # prepares units registry
     app[UnitRegistry.__name__] = UnitRegistry()
